refactor: extract peporag_eval package and unify RAG eval CLI
- add backend/peporag_eval with paths, rag_normalize, ollama_benchmark, rag_eval_runner
- replace duplicated normalize_qwen logic with shared normalize_payload
- wire eval_rag_quality.py to argparse --mode full|raw-qwen; shim eval_rag_quality_raw_qwen.py
- thin benchmark_models.py to delegate to peporag_eval.ollama_benchmark.main
- add module docstrings to backend/scripts/*.py and header comments to *.sh
- configure ruff per-file-ignores E402 for scripts after sys.path bootstrap
- update DEVLOG, qwen_normalization_explainer.ipynb for new entrypoints
