+++ backend/peporag_eval/__init__.py
@@ -1,0 +1,1 @@
+"""PepoRAG offline tooling: Ollama benchmarks and RAG quality evaluation."""
