+++ backend/pyproject.toml
@@ -31,6 +31,9 @@ target-version = "py312"
 select = ["E", "F", "I", "B", "C4", "UP"]
 ignore = []
 
+[tool.ruff.lint.per-file-ignores]
+"scripts/*.py" = ["E402"]
+
 [tool.ruff.format]
 quote-style = "double"
 indent-style = "space"
